# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Flight Data Management API
  description: REST API for managing flight data and integrating CrazySupplier search
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
    description: Local Development Server

paths:
  /flights:
    post:
      summary: Create a new flight
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FlightRequest'
      responses:
        '200':
          description: Flight created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Flight'

    get:
      summary: Search flights (includes CrazySupplier data)
      parameters:
        - in: query
          name: from
          required: true
          schema:
            type: string
            example: JFK
        - in: query
          name: to
          required: true
          schema:
            type: string
            example: LHR
        - in: query
          name: airline
          schema:
            type: string
        - in: query
          name: departureTime
          schema:
            type: string
            format: date-time
            example: 2025-06-10T14:00:00Z
        - in: query
          name: arrivalTime
          schema:
            type: string
            format: date-time
            example: 2025-06-10T22:00:00Z
      responses:
        '200':
          description: Matching flights returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Flight'

  /flights/{id}:
    put:
      summary: Update an existing flight
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FlightRequest'
      responses:
        '200':
          description: Flight updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Flight'

    delete:
      summary: Delete a flight
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Flight deleted successfully

components:
  schemas:
    FlightRequest:
      type: object
      required:
        - airline
        - supplier
        - fare
        - departureAirport
        - destinationAirport
        - departureTime
        - arrivalTime
      properties:
        airline:
          type: string
          example: Delta Airlines
        supplier:
          type: string
          example: LocalSupplier
        fare:
          type: number
          format: double
          example: 499.99
        departureAirport:
          type: string
          example: JFK
        destinationAirport:
          type: string
          example: LHR
        departureTime:
          type: string
          format: date-time
          example: 2025-06-10T14:00:00Z
        arrivalTime:
          type: string
          format: date-time
          example: 2025-06-10T22:00:00Z

    Flight:
      allOf:
        - $ref: '#/components/schemas/FlightRequest'
        - type: object
          properties:
            id:
              type: integer
              example: 1